+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.3
+* Fix Bug var _primarySwatch = await getPrimarySwatch();
+* new option for change color and Brightness at the same time
+
 ## 1.0.2
 * Add DynamicColors Class.